lmisite/requirements.txt
django
django-solo
gunicorn
mysqlclient
django-phonenumber-field
Pillow
django-admin-sortable2
graphene-django>=2.0
pytz
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dateutil
django-ckeditor
django-tinymce
django-filter
sentry-sdk
phonenumbers
beautifulsoup4
nltk
readtime
stripe
django-magiclink
gtin-validator
django-countries
django-mathfilters
